APIF-100 Swagger in app. 'Get Customer Information By CustomerId'. Fixed the failing tests. Added few tests.
To do this. I had to change the domain object as well. I renamed the columns and reordered the properties so that they would match the APIF-104 swagger documentation.

Also reordered the every single line where those properties were used, so that it would be easier to read.

Capitalized the test names, where it was forgotten.
Changed few of the test names to match the checked domain object properties names.
Added few missing tests for checking of property exists.